Creatomg two providers to read details
Dividing the 'packageDetailsProvider.js' in two providers:
'gitPackageDetailsProvider.js' and 'svnPackageDetailsProvider.js' to
read the info  from bower.json from svn.
