+++ lib/service/packageDetailsProvider.js
@@ -3,31 +3,20 @@ var fs = require('fs');
 var path = require('path');
 var Promise = require('bluebird');
 
+var GitPackageDetailsProvider = require('./gitPackageDetailsProvider');
+var SvnPackageDetailsProvider = require('./svnPackageDetailsProvider');
+
 var utils = require('../infrastructure/utils');
 
 module.exports = function PackageDetailsProvider() {
-    var tempFolder = path.join(utils.dirname, 'temp/packageDetails');
-    
+
     function _getPackageDetails(packageUrl) {
-        return new Promise(function(resolve, reject) {
-            var tempName = utils.getRandomString();
-            var gitCloneFolder = path.join(tempFolder, tempName);
-            
-            utils.exec('git clone {0} {1} --depth=1'.format(packageUrl, gitCloneFolder))
-                .then(function() {
-                    var bowerJsonLocation = path.join(gitCloneFolder, 'bower.json');
-                    
-                    var fileContent = fs.readFileSync(bowerJsonLocation);
-                    var bowerJson = JSON.parse(fileContent);
-                    
-                    utils.removeDirectory(gitCloneFolder);
-                    
-                    resolve(bowerJson);
-                })
-                .catch(reject);
-        });
+        if(packageUrl.substring(0, 4) === 'svn+') {
+            return SvnPackageDetailsProvider.getPackageDetails(packageUrl);
+        }
+        return GitPackageDetailsProvider.getPackageDetails(packageUrl);
     }
-    
+
     return {
         getPackageDetails: _getPackageDetails
     };
